Improves pureapi.client
Fixes a bug in get_all_changes() that never yielded the last page of results.
One reason we did not find this bug until now is that the Pure API allows
requesting pages of changes only by resumption token, forcing us to submit
requests sequentially. Because there are often a huge number of change
records every day, previously we avoided requesting all changes in our
integration tests, to reduce the total runtime of automated tests.

We now never give config a default of None in function arguments, but
instead use the Config() constructor. We realized this was safe and
would avoid the gotchas associated with using mutable collections as
argument defaults, because Config() always returns immutable instances.